Remove sharding code from bloom gateway
The sharding by fingerprint range is done on the client side, not on the
server side.

The sharding related code was not used in the gateway.
